# Chat Agent backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Bedrock Chat Configurator API
  description: |
    API backend for the Bedrock Chat Configurator (Flintstone).
    Manages AI assistants, knowledge bases, widget configuration,
    team management, escalation, and billing.

    ## Authentication
    Most endpoints require a **Cognito JWT** token in the `Authorization: Bearer <token>` header.
    Public widget endpoints use an **API key** in the `X-Api-Key` header instead.

    ## Roles
    Role hierarchy (ascending): `viewer` < `editor` < `admin` < `owner`.
    Each endpoint specifies the minimum role required.
  version: 1.0.0
servers:
  - url: https://{apiId}.execute-api.{region}.amazonaws.com/{stage}
    description: AWS API Gateway
    variables:
      apiId:
        default: "2p595psdt1"
      region:
        default: us-west-2
      stage:
        default: dev
        enum: [dev, prod]
...
security:
  - CognitoJWT: []
components:
  securitySchemes:
    CognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito ID token
    ApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Assistant API key for public widget endpoints
  parameters:
    OrganizationId:
      name: X-Organization-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Active organization ID (auto-resolved if user belongs to one org)
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Assistant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, provisioning, ready, error]
        modelConfig:
          $ref: "#/components/schemas/ModelConfig"
        widgetConfig:
          $ref: "#/components/schemas/WidgetConfig"
        apiKey:
          type: string
        allowedDomains:
          type: array
          items:
            type: string
        bedrockAgentId:
          type: string
        bedrockAgentAliasId:
          type: string
        bedrockKnowledgeBaseId:
          type: string
        bedrockGuardrailId:
          type: string
        bedrockGuardrailVersion:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /assistants:
    get:
      tags: [Assistants]
      summary: List all assistants
      description: Returns all assistants for the tenant. Requires viewer+.
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: List of assistants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assistant"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Assistants]
      summary: Create assistant
      description: Creates a new AI assistant with default config. Requires admin+.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Assistant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /assistants/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Assistants]
      summary: Get assistant
      description: Requires viewer+.
      responses:
        "200":
          description: Assistant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Assistants]
      summary: Update assistant
      description: Partial update. Requires editor+.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Assistant"
      responses:
        "200":
          description: Updated assistant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
    delete:
      tags: [Assistants]
      summary: Delete assistant
      description: Requires admin+.
      responses:
        "204":
          description: Deleted
...
  /widget/chat:
    post:
      tags: [Widget]
      summary: Widget chat message
      description: Public endpoint for the embedded widget. Uses API key auth.
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                sessionId:
                  type: string
                userId:
                  type: string
                nodeId:
                  type: string
                image:
                  type: string
                  description: Base64-encoded image
                context:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "200":
          description: Chat reply
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      reply:
                        type: string
                      sessionId:
                        type: string
  /widget/escalate:
    post:
      tags: [Widget]
      summary: Escalate chat to Salesforce
      description: Creates a Salesforce case with the chat transcript. Uses API key auth.
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chatHistory]
              properties:
                chatHistory:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      content:
                        type: string
                      timestamp:
                        type: string
                sessionId:
                  type: string
                context:
                  type: object
                  additionalProperties:
                    type: string
                userInfo:
                  type: object
                  properties:
                    name:
                      type: string
                    email:
                      type: string
                    phone:
                      type: string
                reason:
                  type: string
                attachmentIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Escalation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  caseId:
                    type: string
                  caseNumber:
                    type: string
  /widget/check-escalation:
    post:
      tags: [Widget]
      summary: Check auto-escalation triggers
      description: Analyzes recent messages for escalation triggers. Uses API key auth.
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      content:
                        type: string
                turnCount:
                  type: integer
      responses:
        "200":
          description: Escalation check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  shouldEscalate:
                    type: boolean
                  reason:
                    type: string
  /widget/attachment-url:
    post:
      tags: [Widget]
      summary: Get upload URL for attachment
      description: Returns a presigned URL for uploading files to the HIPAA bucket. Uses API key auth.
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName, contentType]
              properties:
                fileName:
                  type: string
                contentType:
                  type: string
                escalationId:
                  type: string
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      attachmentId:
                        type: string
                      uploadUrl:
                        type: string
                      s3Key:
                        type: string
  /widget/attachment-confirm:
    post:
      tags: [Widget]
      summary: Confirm attachment upload
      description: Confirms upload and triggers EXIF stripping. Uses API key auth.
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attachmentId]
              properties:
                attachmentId:
                  type: string
      responses:
        "200":
          description: Confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      attachmentId:
                        type: string
                      status:
                        type: string

  # ── Knowledge Base Definitions ─────────────────────────────────
  /knowledge-bases:
    get:
      tags: [Knowledge Base Definitions]
      summary: List knowledge base definitions
      description: Returns all KB definitions for the tenant with enriched counts. Requires viewer+.
      responses:
        "200":
          description: List of KB definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeBaseDefinition"
    post:
      tags: [Knowledge Base Definitions]
      summary: Create knowledge base definition
      description: Requires admin+.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                isDefault:
                  type: boolean
      responses:
        "201":
          description: KB definition created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseDefinition"
  /knowledge-bases/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Knowledge Base Definitions]
      summary: Get knowledge base definition
      description: Returns KB definition with linked assistant IDs. Requires viewer+.
      responses:
        "200":
          description: KB definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseDefinition"
    put:
      tags: [Knowledge Base Definitions]
      summary: Update knowledge base definition
      description: Requires editor+.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseDefinition"
    delete:
      tags: [Knowledge Base Definitions]
      summary: Delete knowledge base definition
      description: Fails if assistants are still linked. Requires admin+.
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /guardrails:
    post:
      tags: [Guardrails]
      summary: Create guardrail
      description: Creates a Bedrock guardrail. Requires editor+.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Guardrail created
          content:
            application/json:
              schema:
                type: object
  /guardrails/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    put:
      tags: [Guardrails]
      summary: Update guardrail
      description: Requires editor+.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /attachments:
    get:
      tags: [Attachments]
      summary: List attachments
      description: Requires admin+.
      parameters:
        - name: escalationId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attachments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Attachment"
...
  /widget-presets:
    get:
      tags: [Widget Presets]
      summary: List widget presets
      description: Returns all presets for the tenant. Requires viewer+.
      responses:
        "200":
          description: Presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WidgetPreset"
    post:
      tags: [Widget Presets]
      summary: Create widget preset
      description: Requires editor+.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, primaryColor]
              properties:
                name:
                  type: string
                position:
                  type: string
                primaryColor:
                  type: string
                secondaryColor:
                  type: string
                customLauncherIconUrl:
                  type: string
                customLauncherHtml:
                  type: string
                customCss:
                  type: string
                typingIndicatorStyle:
                  type: string
                typingPhrases:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Preset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetPreset"
  /widget-presets/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Widget Presets]
      summary: Get widget preset
      description: Requires viewer+.
      responses:
        "200":
          description: Preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetPreset"
    put:
      tags: [Widget Presets]
      summary: Update widget preset
      description: Requires editor+.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetPreset"
      responses:
        "200":
          description: Updated preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetPreset"
    delete:
      tags: [Widget Presets]
      summary: Delete widget preset
      description: Requires editor+.
      responses:
        "204":
          description: Deleted